+++ src/DisjointDatabase.cpp
@@ -5,7 +5,7 @@
 DisjointDatabase::DisjointDatabase(
     int len, std::string name, std::vector<std::vector<std::vector<int>>> grids)
     : where(len, -1) {
-    for (int i = 0; i < grids.size(); i++) {
+    for (size_t i = 0; i < grids.size(); i++) {
         auto pd = std::make_shared<PartialDatabase>(grids[i], name, i);
         databases.push_back(pd);
 
@@ -41,7 +41,7 @@ int DisjointDatabase::numPatterns() const { return databases.size(); }
 int DisjointDatabase::getHeuristic(
     const std::vector<uint64_t>& positions) const {
     int totalDist = 0;
-    for (int i = 0; i < positions.size(); i++) {
+    for (size_t i = 0; i < positions.size(); i++) {
         totalDist += databases[i]->distMap[positions[i]];
     }
     return totalDist;
